# libkmod: Improve index dump performance