--- a/libkmod/libkmod-index.h
+++ b/libkmod/libkmod-index.h
@@ -19,7 +19,7 @@ struct index_file;
 struct index_file *index_file_open(const char *filename);
 void index_file_close(struct index_file *idx);
 char *index_search(struct index_file *idx, const char *key);
-void index_dump(struct index_file *in, int fd, bool alias_prefix);
+void index_dump(struct index_file *in, FILE *fp, bool alias_prefix);
 struct index_value *index_searchwild(struct index_file *idx, const char *key);
 
 void index_values_free(struct index_value *values);
@@ -31,4 +31,4 @@ int index_mm_open(const struct kmod_ctx *ctx, const char *filename,
 void index_mm_close(struct index_mm *index);
 char *index_mm_search(const struct index_mm *idx, const char *key);
 struct index_value *index_mm_searchwild(const struct index_mm *idx, const char *key);
-void index_mm_dump(const struct index_mm *idx, int fd, bool alias_prefix);
+void index_mm_dump(const struct index_mm *idx, FILE *fp, bool alias_prefix);