--- a/shared/util.c
+++ b/shared/util.c
@@ -237,32 +237,6 @@ ssize_t read_str_safe(int fd, char *buf, size_t buflen)
 	return done;
 }
 
-ssize_t write_str_safe(int fd, const char *buf, size_t buflen)
-{
-	size_t todo = buflen;
-	size_t done = 0;
-
-	assert_cc(EAGAIN == EWOULDBLOCK);
-
-	do {
-		ssize_t r = write(fd, buf + done, todo);
-
-		if (r == 0)
-			break;
-		else if (r > 0) {
-			todo -= r;
-			done += r;
-		} else {
-			if (errno == EAGAIN || errno == EINTR)
-				continue;
-			else
-				return -errno;
-		}
-	} while (todo > 0);
-
-	return done;
-}
-
 int read_str_long(int fd, long *value, int base)
 {
 	char buf[32], *end;