@stoeckmann 2024-11-23
Will close this one since we prefer https://github.com/kmod-project/kmod/pull/252.